bugs fixed

+++ main.js
@@ -185,6 +185,7 @@ function SVGEditor(DOM, options) {
     });
 
     document.addEventListener('keydown', function(event) {
+        event.preventDefault();
         if(event.keyCode ==90 && event.metaKey) {
             console.log("Відміна");
         } else if (event.keyCode === 224) {
@@ -194,6 +195,7 @@ function SVGEditor(DOM, options) {
     });
 
     document.addEventListener('keyup', function(event){
+        event.preventDefault();
         if(event.keyCode === 224 && polygon) {
             polygon.resumeDrawing();
             var coords = {
@@ -341,14 +343,17 @@ function Polygon(parentElem) {
 
     this.stopDraw = function() {
         var polygonPoints = "";
-        polygonCoords.pop();
+        if(!editable)
+            polygonCoords.pop();
         polygonCoords.forEach(elem => {
             polygonPoints += elem.x + ',' + elem.y  + ' ';
         });
 
         polygon.setAttribute('points', polygonPoints);
         polygon.style.opacity = '0.4';
-        dot.delete();
+        
+        if(!editable)
+            dot.delete();
     }
 
     this.remove = function() {
